requirements.txt
amqp==5.3.1
asgiref==3.11.0
autoflake==2.3.1
billiard==4.2.3
black==25.11.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
flake8==7.3.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==5.1.0
pytesseract==0.3.13
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
ruff==0.14.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
